@@ -11,17 +11,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.15",
-    "@angular/common": "^7.2.15",
-    "@angular/compiler": "^7.2.15",
-    "@angular/core": "^7.2.15",
-    "@angular/forms": "^7.2.15",
-    "@angular/http": "^7.2.15",
-    "@angular/platform-browser": "^7.2.15",
-    "@angular/platform-browser-dynamic": "^7.2.15",
-    "@angular/router": "^7.2.15",
+    "@angular/animations": "^8.2.12",
+    "@angular/common": "^8.2.12",
+    "@angular/compiler": "^8.2.12",
+    "@angular/core": "^8.2.12",
+    "@angular/forms": "^8.2.12",
+    "@angular/platform-browser": "^8.2.12",
+    "@angular/platform-browser-dynamic": "^8.2.12",
+    "@angular/router": "^8.2.12",
     "@fortawesome/fontawesome-free": "5.4.1",
-    "@ng-bootstrap/ng-bootstrap": "3.3.1",
     "aws-sdk": "2.452.0",
     "bcrypt": "3.0.2",
     "body-parser": "1.18.3",
@@ -41,26 +39,26 @@
     "multer": "1.4.1",
     "mysql2": "1.6.1",
     "ngx-cookie-service": "1.0.10",
-    "ngx-file-drop": "6.0.1",
-    "ngx-spinner": "6.1.2",
+    "ngx-file-drop": "8.0.8",
+    "ngx-spinner": "8.0.3",
     "nodemailer": "4.6.8",
     "popper.js": "1.14.4",
-    "rxjs": "^6.5.2",
+    "rxjs": "^6.5.3",
     "sequelize": "5.15.1",
     "tslib": "^1.9.0",
-    "zone.js": "~0.8.29"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.13.9",
-    "@angular/cli": "7.3.9",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/language-service": "^7.2.15",
+    "@angular-devkit/build-angular": "~0.803.15",
+    "@angular/cli": "8.3.15",
+    "@angular/compiler-cli": "^8.2.12",
+    "@angular/language-service": "^8.2.12",
     "@types/bootstrap": "4.1.2",
     "@types/jasmine": "~2.8.9",
     "@types/jasminewd2": "~2.0.5",
     "@types/jquery": "3.3.22",
     "@types/node": "~10.12.0",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.0.1",
     "eslint": "5.16.0",
     "eslint-config-standard": "12.0.0",
     "eslint-plugin-import": "2.17.2",
@@ -77,6 +75,6 @@
     "protractor": "~5.4.1",
     "ts-node": "~7.0.1",
     "tslint": "~5.11.0",
-    "typescript": "~3.2.4"
+    "typescript": "~3.5.3"
   }
 }
